fix: :pencil2: rename general interface to fix typo
